Saving data to sqlite database.
