+++ node_server/database/helper/__init__.py
@@ -7,10 +7,9 @@ class Database:
     __DATABASE_LOCATION = os.path.join(os.path.dirname(__file__), '..', 'blockchain.db')
 
     def __init__(self):
-        print(self.__DATABASE_LOCATION, flush=True)
         self.connection = sqlite3.connect(self.__DATABASE_LOCATION)
 
-    def execute(self,sql):
+    def execute(self,sql, object=None):
         cursor = self.connection.cursor()
         cursor.execute(sql)
         self.connection.commit()
