+++ node_server/models/Blockchain/Block/Transaction/Content/__init__.py
@@ -1,3 +1,4 @@
+from database.helper import Database
 class Content():
 
 
@@ -9,3 +10,11 @@ class Content():
     @classmethod
     def from_json(cls, data: dict):
         return cls(**data)
+
+    def save(self, transaction_id):
+        sql = '''
+        INSERT INTO content (text, name, email, id_transaction)
+        VALUES ("{0}","{1}","{2}",{3})
+        '''.format(self.name, self.email, self.text, transaction_id)
+        database = Database()
+        database.execute(sql)
