+++ node_server/models/Blockchain/__init__.py
@@ -43,6 +43,8 @@ class Blockchain:
 
         block.hash = proof
         self.chain.append(block)
+        for transaction in block.transactions:
+            transaction.save()
         return True
 
     @staticmethod
